+++ src/headless/plugins/roster/plugin.js
@@ -38,6 +38,7 @@ converse.plugins.add('converse-roster', {
 
         const { __ } = _converse;
         const labels = {
+            HEADER_UNSAVED_CONTACTS: __('Unsaved contacts'),
             HEADER_CURRENT_CONTACTS: __('My contacts'),
             HEADER_PENDING_CONTACTS: __('Pending contacts'),
             HEADER_REQUESTING_CONTACTS: __('Contact requests'),