+++ src/headless/types/plugins/roster/types.d.ts
@@ -1,0 +1,9 @@
+export type RosterContactAttributes = {
+    jid: string;
+    subscription: ('none' | 'to' | 'from' | 'both');
+    ask?: 'subscribe';
+    name?: string;
+    groups?: string[];
+    requesting?: boolean;
+};
+//# sourceMappingURL=types.d.ts.map