+++ src/plugins/rosterview/types.ts
@@ -1,0 +1,5 @@
+import RosterContact from "@converse/headless/types/plugins/roster/contact"
+
+export type ContactsMap = {
+    [Key: string]: RosterContact[]
+}